+++ src/error/mod.rs
@@ -1,5 +1,8 @@
+pub use evalexpr::EvalexprError;
+
 pub use read::ReadBinError;
 pub use write::WriteBinError;
 
 mod read;
 mod write;
+